New bundled assembly mapping code
`Xamarin.Android` stores uncompressed assemblies inside the APK archive in order
to be able to map them to the device's memory and hand the pointer to Mono,
without having to unpack the files from the APK. We've been using an embedded
copy of the `unzip` utility in order to access the archive. Not only the code
was old, it also did much more than was necessary for our purposes. This
incurred runtime overhead which this commit aims to fix by doing the following:

  * Remove the old C code and replace with a simple implementation tailored
    specifically to our needs

  * Delay mapping of assemblies, debug information files and config files until
    such time when Mono asks us to load a required assembly - this should
	improve application startup time.

  * Store information about the bundled assemblies in a dynamically allocated
    array without constantly reallocating it (and thus copying data) at the
	expense of a little memory wasted (up to 16k currently)

In addition to the above performance changes, the following C++ features are
added here:

  * Automatic memory management (via RAII) for memory allocated from the heap
    but used only within a single function (the `simple_pointer_guard` class)

  * Implement portions of the C++ standard header `type_traits` (which is
    shipped with the NDK but for some reason not found by `clang++`)

Performance impact with our Xamarin.Forms integration test, Android 10 on Pixel
3 XL. Measurements are in milliseconds, Xamarin.Android total init time only:

|  **Assembly preload** |  **master** | **this commit** | **Architecture** |
|-----------------------|------------ |-----------------|------------------|
| **enabled**           | 271.373     | 216.931         | armeabi-v7a      |
| **disabled**          | 238.404     | 181.080         | armeabi-v7a      |
| **enabled**           | 255.984     | 201.498         | arm64-v8a        |
| **disabled**          | 225.933     | 169.507         | arm64-v8a        |