[Mono.Android] Delegate.CreateDelegate & generic type definitions (#4398)
Fixes: https://github.com/xamarin/xamarin-android/issues/4394

Context: https://github.com/xamarin/xamarin-android/commits/test-net5-drop

When using a preview drop of the .NET 5 BCL with a Xamarin.Forms app,
app startup crashes within
`AndroidTypeManager.RegisterNativeMembers()` because we attempt to use
`Delegate.CreateDelegate()` on a generic type definition, a
`typeof(VisualElementRenderer<>)`.

[.NET Core and .NET Framework doesn't like this][0]; [mono does][1],
which is why it hasn't been a problem before.

We attempt to register the generic type definition for
`VisualElementRenderer<T>` because that's what the Java Callable
Wrapper does!  Given C#:

	// C#
	partial class FormsViewGroup : Android.Views.ViewGroup {
	}
	partial class VisualElementRenderer<T> : FormsViewGroup {
	}
	partial class PageRenderer : VisualElementRenderer<Page> {
	}

the Java Callable Wrappers will "mirror" that hierarchy, with
`Runtime.register()` calls in the static initializers:

	// Java Callable Wrappers
	public class FormsViewGroup extends android.view.ViewGroup {}
	public class VisualElementRenderer_1 extends FormsViewGroup {
	  static {
	    mono.android.Runtime.register ("VisualElementRenderer`1, …");
	  }
	}
	public class PageRenderer extends VisualElementRenderer_1 {
	  static {
	    mono.android.Runtime.register ("PageRenderer, …");
	  }
	}

When a `PageRenderer` instance is created from C#, an instance of the
Java Callable Wrapper `PageRenderer` will also be created, but before
the instance an be constructed the static initializer of
`PageRenderer` and all base classes must be executed *first*, and
because there's a `Runtime.register()` for
``VisualElementRenderer`1``, `RegisterNativeMembers()` will thus
attempt to register the generic type definition.

How should this be fixed?

Assuming that updating .NET Core to behave like Mono isn't an option,
there are three other solutions:

 1. Don't emit `Runtime.register()` in the Java Callable Wrappers for
    C# types which are generic.

 2. Within `AndroidRuntime.RegisterNativeMembers()`, skip generic type
    definitions entirely.

 3. Improve the semantics of connector method lookups.

(1) and (2) are in fact [not viable solutions][2], because the
generic type may override Java methods that *must* be registered, and
not registering them will result in `Java.Lang.UnsatisfiedLinkError`s
when invoking the method from Java.

`ExceptionTests.ManagedJavaManaged_FinallyExecuted()` has been
updated to demonstrate this scenario.

The whole reason that we can lookup connector methods on a generic
type definition in the first place -- it works on mono! -- is because
the connector methods are in fact present in a *non*-generic base
class, and Mono looks at all base classes to resolve a method before
deciding to throw.

Instead of relying on a Mono feature, do this ourselves.

[0]: https://github.com/dotnet/runtime/issues/33531
[1]: https://github.com/mono/mono/issues/19207
[2]: https://github.com/xamarin/xamarin-android/issues/4394#issuecomment-599082102
[3]: https://github.com/xamarin/xamarin-android/blob/8599d63d28922a580c75aeb015554ae9d5034bf3/src/Mono.Android/Test/Java.Interop/JnienvTest.cs#L468-L472